Move read/write object tests to test_read_write_object module
