+++ test/on_yubikey/cli_piv/test_read_write_object.py
@@ -22,6 +22,20 @@ def additional_tests(ykman_cli):
             ykman_cli('piv', 'reset', '-f')
             pass
 
+        def test_write_read_preserves_ansi_escapes(self):
+            red = b'\x00\x1b[31m'
+            blue = b'\x00\x1b[34m'
+            reset = b'\x00\x1b[0m'
+            data = (b'Hello, ' + red + b'red' + reset + b' and ' + blue
+                    + b'blue' + reset + b' world!')
+            ykman_cli(
+                'piv', 'write-object',
+                '-m', DEFAULT_MANAGEMENT_KEY, '0x5f0001',
+                '-', input=data)
+            output_data = ykman_cli.with_bytes_output(
+                'piv', 'read-object', '0x5f0001')
+            self.assertEqual(data, output_data)
+
         def test_read_write_read_is_noop(self):
             data = os.urandom(32)
